Updates slider image component tests
